Read JSON from online source into List
First, the UserList class was deemed unnecessary, so it was removed.

Second, the friends array declaration was uncommented because a Friend
struct exists now.

Finally, the capability for downloading and decoding JSON data from
a URL was added to ContentView. This is an asynchronous function. Note
that the date decoding strategy ".iso8601" is used so that the date in
the JSON may be parsed properly.

All changes are currently working properly; the list displays and there
seem to be no errors.