Create and configure Core Data model and entities
A Core Data model was created, containing three entities: CachedUser is
the local version of User, CachedFriend is the local version of Friend,
and TagHolder is the local version of one tag string within the array
for an instance of User.

CachedUser has a one-to-many relationship with CachedFriend, and
CachedUser has a one-to-many relationship with TagHolder.

Finally, both CachedUser and CachedFriend have a constraint on the id
attribute.